Preclear Disk version 2020.01.16
- Fix: plugin tries to gently stop the preclear script
- Fix: addBannerWarning retrocompatibility
- Add: sorting on Tools > Preclear for Preclear Queue
- Add: Pause/Stop/Resume all
- Fix: script: version 1.0.8
- Fix: script: resume not working while verifying the disk
- Fix: script: correctly compute current speed even if the session is paused/enqueued
- Fix: script: increase ui refresh rate
- Fix: script: better exit handling

+++ source/preclear.disk/script/preclear_queue.sh
@@ -12,7 +12,7 @@ debug() {
 exec 2> >(while read err; do echo "$(date +"%b %d %T" ) preclear_queue: ${err}" >> /var/log/preclear.disk.log; echo "${err}"; done; >&2)
 
 get_running_sessions() {
-  for file in $(ls -tr /tmp/.preclear/*/pid 2>/dev/null); do
+  for file in $(ls /tmp/.preclear/*/pid 2>/dev/null); do
     pid=$(cat $file);
     if [ -e "/proc/${pid}/exe" ]; then
       echo $(echo $file | cut -d'/' -f4);
@@ -34,6 +34,32 @@ do_clean()
   tmux kill-window -t 'preclear_queue' 2>/dev/null
 }
 
+sort_running()
+{
+  local sort_file="/boot/config/plugins/preclear.disk/sort_order"
+  local  -g -a sort_order
+  local i=999999
+  if [ -f "$sort_file" ]; then
+    for disk in $(get_running_sessions); do
+      local line=""
+      if [ -f "$sort_file" ]; then
+        line=$(awk "/$disk/{ print NR; exit }" $sort_file)
+      fi
+      if [ -n "$line" ]; then
+        sort_order[$line]=$disk
+      else
+        sort_order[$i]=$disk
+        i=$((i+1))
+      fi
+    done
+    for disk in "${sort_order[@]}"; do
+      echo $disk
+    done
+  else
+    get_running_sessions
+  fi
+}
+
 queue=${1-1};
 timer=$(date '+%s')
 
@@ -44,10 +70,11 @@ trap "do_clean;" exit
 
 while [ -f /var/run/preclear_queue.pid ]; do
   i=0
-  for disk in $(get_running_sessions); do
+  for disk in $(sort_running); do
     tmpdir="/tmp/.preclear/${disk}"
     queued="${tmpdir}/queued"
-    if [ -d $tmpdir ]; then
+    paused="${tmpdir}/pause"
+    if [ -d $tmpdir -a ! -f $paused ]; then
       if [ $i -lt $queue -a -f $queued ]; then
         debug "Restoring $disk preclear session"
         rm $queued
@@ -55,8 +82,8 @@ while [ -f /var/run/preclear_queue.pid ]; do
         debug "Enqueuing $disk preclear session"
         touch $queued
       fi
+      i=$(( $i + 1 ))
     fi
-    i=$(( $i + 1 ))
     timer=$(date '+%s')
   done
   if [ "$i" -eq "0" ] && [[  $(( $(date '+%s') - $timer )) -gt 60 ]]; then
